refs #127: ログインレイアウトの作成
