+++ src/pages/index.tsx
@@ -1,10 +1,8 @@
-import { useAuth } from '@/context/auth';
+import UnauthenticatedLayout from '@/layout/auth/UnauthenticatedLayout';
 import { login } from '@/lib/auth';
 import { css } from '@emotion/react';
 import Image from 'next/image';
 import Link from 'next/link';
-import { useRouter } from 'next/router';
-import { useEffect } from 'react';
 
 const styles = {
   container: css`
@@ -52,47 +50,39 @@ const styles = {
 };
 
 export default function Home() {
-  const { fbUser, isLoading } = useAuth();
-  const router = useRouter();
-
-  useEffect(() => {
-    if (fbUser) {
-      router.push('/project-list');
-    }
-  }, [fbUser, isLoading, router]);
-
-  if (fbUser || isLoading) {
-    return null;
-  }
-
   return (
-    <div css={styles.container}>
-      <div css={styles.titleWrapper}>
-        <h1>GitHubの直近1年間のデータ閲覧サイト</h1>
-      </div>
-      <div css={styles.main}>
-        <Image
-          src={'/images/main.png'}
-          alt={'グラフ'}
-          height={380}
-          width={480}
-          css={styles.image}
-        />
-        <div css={styles.discWrapper}>
-          <p>
-            GitHubアカウントでログインすると、直近1年間のデータを閲覧できたり、他のユーザーと比較できたりします。
-          </p>
-          <button css={styles.loginButton} onClick={login}>
-            ログイン
-          </button>
+    <UnauthenticatedLayout>
+      <div css={styles.container}>
+        <div css={styles.titleWrapper}>
+          <h1>GitHubの直近1年間のデータ閲覧サイト</h1>
         </div>
+        <div css={styles.main}>
+          <Image
+            src={'/images/main.png'}
+            alt={'グラフ'}
+            height={380}
+            width={480}
+            css={styles.image}
+          />
+          <div css={styles.discWrapper}>
+            <p>
+              GitHubアカウントでログインすると、直近1年間のデータを閲覧できたり、他のユーザーと比較できたりします。
+            </p>
+            <button
+              css={styles.loginButton}
+              onClick={login}
+            >
+              ログイン
+            </button>
+          </div>
+        </div>
+        <p>
+          ※以前の仕様を変更しているため以前までのページは移動しました。
+          <br />
+          現在データを書き換えいるため、以前のユーザーの表示を減らしています。
+        </p>
+        <Link href="/old">以前のページはこちら</Link>
       </div>
-      <p>
-        ※以前の仕様を変更しているため以前までのページは移動しました。
-        <br />
-        現在データを書き換えいるため、以前のユーザーの表示を減らしています。
-      </p>
-      <Link href="/old">以前のページはこちら</Link>
-    </div>
+    </UnauthenticatedLayout>
   );
 }
