+++ src/pages/project-list.tsx
@@ -4,7 +4,6 @@ import Sidebar from '@/components/Sidebar';
 import Topbar from '@/components/Topbar';
 import { useAuth } from '@/context/auth';
 import { css } from '@emotion/react';
-import { useRouter } from 'next/router';
 import { useEffect } from 'react';
 import { getProjectHandler } from '@/lib/firebase/getProjectHandler';
 import { ProjectType } from '@/types/ProjectType';
@@ -12,6 +11,7 @@ import { fetchProjectsState } from '@/store/fetchProjectAtoms';
 import { useRecoilState } from 'recoil';
 import SidebarTitle from '@/components/Sidebar/SidebarTitle';
 import { isOpenSidebarState } from '@/store/sidebarAtoms';
+import AuthenticatedLayout from '@/layout/auth/AuthenticatedLayout';
 
 const styles = {
   container: css`
@@ -44,8 +44,7 @@ const styles = {
 };
 
 const ProjectList = () => {
-  const { fbUser, isLoading } = useAuth();
-  const router = useRouter();
+  const { fbUser } = useAuth();
   const [fetchProjects, setFetchProjects] = useRecoilState<
     ProjectType[]
   >(fetchProjectsState);
@@ -53,12 +52,6 @@ const ProjectList = () => {
   const [isOpenSidebar, setIsOpenSidebar] = useRecoilState(
     isOpenSidebarState
   );
-
-  useEffect(() => {
-    if (!isLoading && !fbUser) {
-      router.push('/');
-    }
-  }, [fbUser, isLoading, router]);
 
   useEffect(() => {
     if (fbUser) {
@@ -78,24 +71,22 @@ const ProjectList = () => {
     }
   }, [fbUser]);
 
-  if (!fbUser || isLoading) {
-    return null;
-  }
-
   return (
-    <div css={styles.container}>
-      <div css={styles.topContent}>
-        <SidebarTitle />
-        <Topbar />
-      </div>
-      <div css={styles.mainWrapper}>
-        {isOpenSidebar && <Sidebar />}
-        <div css={styles.projectWrapper}>
-          <CreateProjectBox />
-          <ProjectBox />
+    <AuthenticatedLayout>
+      <div css={styles.container}>
+        <div css={styles.topContent}>
+          <SidebarTitle />
+          <Topbar />
+        </div>
+        <div css={styles.mainWrapper}>
+          {isOpenSidebar && <Sidebar />}
+          <div css={styles.projectWrapper}>
+            <CreateProjectBox />
+            <ProjectBox />
+          </div>
         </div>
       </div>
-    </div>
+    </AuthenticatedLayout>
   );
 };
